Additional safeguards against unauthorized access
In testing, it became apparent that there were some ways to add (but not remove)
emails from lists which were in the same suffix namespace but where the
moderator user lacked sufficient permissions.  In addition, emails could be
added to the unified list if defined in the YAML file, even if they did not
correspond to a valid subgroup.

This patch fixes all of this, and outputs errors when a subgroup can't be
properly accessed.

Finally, this should make it safer to live dangerously by not defining a suffix
at all, and relying solely upon Groups.io permissions to determine whether a
group defined in the YAML file should be permitted to make changes to Groups.io.
This requires more testing, though it would have the advantage of being less
confusing to users because without list_suffix the group name defined in the
YAML would map 1:1 to the subgroup name.

Signed-off-by: Brian Warner <brian@bdwarner.com>