Add error reporting and protection for main list
This commit adds a number of protections against modifying lists which shouldn't
be modified. It also explicitly prevents modifications to the main list, which
would be a terrible idea (not the least being it will wipe out all of the
subscribers who aren't defined via the yaml file).

It also adds error reporting, which can be useful in tracking down issues like
incorrect moderator account permissions.

It is also now safe to define lists without using suffixes.  You can still use
suffixes if you want to define a specific namespace of YAML-managed subgroups,
but it's also safe to do without if you want.

Signed-off-by: Brian Warner <brian@bdwarner.com>